% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ACE_calc.R
\name{ACE_calc}
\alias{ACE_calc}
\title{A function for computing the accumulated cyclone energy of a given storm.}
\usage{
ACE_calc(storm_id, data = hurdat)
}
\arguments{
\item{storm_id}{a character value}

\item{data}{a data frame}
}
\value{
the accumulated storm energy of the chosen storm
}
\description{
This function takes two arguments: a character value representing a storm id and
a data frame where the storm information is located.
}
\examples{
ACE_calc("AL092004", data = hurdat)

}
