% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpolate_30min.R
\name{interpolate_30min}
\alias{interpolate_30min}
\title{A function for interpolating a storm track to 30 minute increments}
\usage{
interpolate_30min(track_id, data = hurdat)
}
\arguments{
\item{data}{a data frame}

\item{storm_id}{a character value}
}
\value{
a data frame with the interpolated data.
}
\description{
This function takes two arguments: a character value representing a storm id and
a data frame where the storm information is located.
}
\examples{
interpolate_30min("AL092004", data = hurdat)

}
