% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_US_landfall.R
\name{is_US_landfall}
\alias{is_US_landfall}
\title{A function for determining whether a storm made landfall in the continental US}
\usage{
is_US_landfall(storm_id, data = hurdat)
}
\arguments{
\item{storm_id}{a character value}

\item{data}{a data frame}
}
\value{
a logical vector indicating whether the storm made landfall or not
}
\description{
This function takes two arguments: a character value representing a storm id and
a data frame where the storm information is located.
}
\examples{
is_US_landfall("AL092004", data = hurdat)

}
