% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotstorms.R
\name{plotstorms}
\alias{plotstorms}
\title{A function for plotting a map of storm tracks for a selection of storms.}
\usage{
plotstorms(id_list, data = hurdat)
}
\arguments{
\item{id_list}{a list of character values representing the storm ids}

\item{data}{a data frame}
}
\value{
a plot of the storm tracks
}
\description{
This function takes two arguments: a list of storm ids and a data frame
where the storm information is located.
}
\examples{
plotstorms(c("AL011851", "AL011852", "AL162003"), data = hurdat)

}
