% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stormsize.R
\name{stormsize}
\alias{stormsize}
\title{A function for plotting a map of a selected storm at one point and arms representing wind traveled in nautical miles from the origin.}
\usage{
stormsize(rownum, dataf = hurdat)
}
\arguments{
\item{rownum}{a row number within the data frame}

\item{dataf}{a data frame}
}
\value{
a plot of the origin and how far the winds have moved in the NE, SE, SW, and NW directions
}
\description{
This function takes two arguments: a row number within the data frame and a data frame
where the storm information is located.
}
\examples{
stormsize(52299, dataf = hurdat)

}
